event/requirements.txt
attrs==22.2.0
Automat==22.10.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==39.0.0
cssselect==1.2.0
filelock==3.9.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.7.0
itemloaders==1.0.6
jmespath==1.0.1
lxml==4.9.2
packaging==23.0
parsel==1.7.0
Protego==0.2.1
psycopg==3.1.7
psycopg-binary==3.1.6
psycopg-pool==3.1.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.6
pyOpenSSL==23.0.0
python-decouple==3.7
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
Scrapy==2.7.1
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.2.post1
tldextract==3.4.0
Twisted==22.10.0
typing_extensions==4.4.0
urllib3==1.26.13
w3lib==2.1.1
zope.interface==5.5.2
